Mock dashboards moved to spec_helper to be shared
