+++ spec/spec_helper.rb
@@ -1,4 +1,8 @@
 require "webmock/rspec"
+require "administrate/fields/string"
+require "administrate/fields/email"
+require "administrate/fields/number"
+require "administrate/base_dashboard"
 
 # http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 RSpec.configure do |config|
@@ -14,3 +18,30 @@ RSpec.configure do |config|
 end
 
 WebMock.disable_net_connect!(allow_localhost: true, allow: "percy.io")
+
+class MockDashboard < Administrate::BaseDashboard
+  ATTRIBUTE_TYPES = {
+    name: Administrate::Field::String,
+    email: Administrate::Field::Email,
+    phone: Administrate::Field::Number,
+  }
+end
+
+class DashboardWithAnArrayOfScopes < Administrate::BaseDashboard
+  ATTRIBUTE_TYPES = {
+    name: Administrate::Field::String,
+  }
+
+  COLLECTION_SCOPES = [:active, :old, "with_argument(3)"]
+end
+
+class DashboardWithAHashOfScopes < Administrate::BaseDashboard
+  ATTRIBUTE_TYPES = {
+    name: Administrate::Field::String,
+  }
+
+  COLLECTION_SCOPES = {
+    status: [:active, :inactive],
+    other: [:last_week, :old, "with_argument(3)"],
+  }
+end
